+++ app/lib/helpers/__tests__/titleCase.test.ts
@@ -1,4 +1,4 @@
-import { titleCase } from '../index';
+import { titleCase } from '../helpers';
 
 describe('parseTimestampToDate', () => {
   it('should throw on empty string', () => {
